requirements.txt
sentence-transformers>=3.0.0
pydantic>=2.0.0
msgpack>=1.0.0
zstandard>=0.23.0
PyNaCl>=1.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
